+++ Source/BugsnagBreadcrumb.m
@@ -49,15 +49,30 @@ NSString *BSGBreadcrumbTypeValue(BSGBreadcrumbType type) {
     }
 }
 
+BSGBreadcrumbType BSGBreadcrumbTypeFromString(NSString *value) {
+    if ([value isEqual:@"log"]) {
+        return BSGBreadcrumbTypeLog;
+    } else if ([value isEqual:@"user"]) {
+        return BSGBreadcrumbTypeUser;
+    } else if ([value isEqual:@"error"]) {
+        return BSGBreadcrumbTypeError;
+    } else if ([value isEqual:@"state"]) {
+        return BSGBreadcrumbTypeState;
+    } else if ([value isEqual:@"process"]) {
+        return BSGBreadcrumbTypeProcess;
+    } else if ([value isEqual:@"request"]) {
+        return BSGBreadcrumbTypeRequest;
+    } else if ([value isEqual:@"navigation"]) {
+        return BSGBreadcrumbTypeNavigation;
+    } else {
+        return BSGBreadcrumbTypeManual;
+    }
+}
+
 @interface BugsnagBreadcrumbs ()
 
 @property(nonatomic, readwrite, strong) NSMutableArray *breadcrumbs;
 @property(nonatomic, readonly, strong) dispatch_queue_t readWriteQueue;
-@end
-
-@interface BugsnagBreadcrumb ()
-
-- (NSDictionary *_Nullable)objectValue;
 @end
 
 @implementation BugsnagBreadcrumb
@@ -170,6 +185,24 @@ NSString *BSGBreadcrumbTypeValue(BSGBreadcrumbType type) {
     return nil;
 }
 
++ (instancetype)breadcrumbFromDict:(NSDictionary *)dict {
+    BOOL isValidCrumb = [dict[BSGKeyType] isKindOfClass:[NSString class]]
+        && [dict[BSGKeyTimestamp] isKindOfClass:[NSString class]]
+        && [dict[BSGKeyMetadata] isKindOfClass:[NSDictionary class]]
+        // Accept legacy 'name' value if provided.
+        && ([dict[BSGKeyMessage] isKindOfClass:[NSString class]]
+            || [dict[BSGKeyName] isKindOfClass:[NSString class]]);
+    if (isValidCrumb) {
+        return [self breadcrumbWithBlock:^(BugsnagBreadcrumb *crumb) {
+            crumb.message = dict[BSGKeyMessage] ?: dict[BSGKeyName];
+            crumb.metadata = dict[BSGKeyMetadata];
+            crumb.timestamp = [[Bugsnag payloadDateFormatter] dateFromString:dict[BSGKeyTimestamp]];
+            crumb.type = BSGBreadcrumbTypeFromString(dict[BSGKeyType]);
+        }];
+    }
+    return nil;
+}
+
 @end
 
 @implementation BugsnagBreadcrumbs
@@ -181,6 +214,7 @@ NSUInteger BreadcrumbsDefaultCapacity = 25;
     if (self = [super init]) {
         _breadcrumbs = [NSMutableArray new];
         _capacity = BreadcrumbsDefaultCapacity;
+        _enabledBreadcrumbTypes = BSGEnabledBreadcrumbTypeAll;
         _readWriteQueue = dispatch_queue_create("com.bugsnag.BreadcrumbRead",
                                                 DISPATCH_QUEUE_SERIAL);
         NSString *cacheDir = [NSSearchPathForDirectoriesInDomains(
@@ -205,7 +239,7 @@ NSUInteger BreadcrumbsDefaultCapacity = 25;
         return;
     }
     BugsnagBreadcrumb *crumb = [BugsnagBreadcrumb breadcrumbWithBlock:block];
-    if (crumb) {
+    if (crumb && [self shouldSaveType:crumb.type]) {
         [self resizeToFitCapacity:self.capacity - 1];
         dispatch_barrier_sync(self.readWriteQueue, ^{
             [self.breadcrumbs addObject:crumb];
@@ -242,6 +276,27 @@ NSUInteger BreadcrumbsDefaultCapacity = 25;
         }
     });
     return [cache isKindOfClass:[NSArray class]] ? cache : nil;
+}
+
+- (BOOL)shouldSaveType:(BSGBreadcrumbType)type {
+    switch (type) {
+        case BSGBreadcrumbTypeManual:
+            return YES;
+        case BSGBreadcrumbTypeError:
+            return self.enabledBreadcrumbTypes & BSGEnabledBreadcrumbTypeError;
+        case BSGBreadcrumbTypeLog:
+            return self.enabledBreadcrumbTypes & BSGEnabledBreadcrumbTypeLog;
+        case BSGBreadcrumbTypeNavigation:
+            return self.enabledBreadcrumbTypes & BSGEnabledBreadcrumbTypeNavigation;
+        case BSGBreadcrumbTypeProcess:
+            return self.enabledBreadcrumbTypes & BSGEnabledBreadcrumbTypeProcess;
+        case BSGBreadcrumbTypeRequest:
+            return self.enabledBreadcrumbTypes & BSGEnabledBreadcrumbTypeRequest;
+        case BSGBreadcrumbTypeState:
+            return self.enabledBreadcrumbTypes & BSGEnabledBreadcrumbTypeState;
+        case BSGBreadcrumbTypeUser:
+            return self.enabledBreadcrumbTypes & BSGEnabledBreadcrumbTypeUser;
+    }
 }
 
 @synthesize capacity = _capacity;