+++ Source/BugsnagNotifier.m
@@ -347,7 +347,6 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
                       onCrash:&BSSerializeDataCrashHandler];
     [self computeDidCrashLastLaunch];
     [self setupConnectivityListener];
-    [self updateAutomaticBreadcrumbDetectionSettings];
 
     NSNotificationCenter *center = [NSNotificationCenter defaultCenter];
     [self watchLifecycleEvents:center];
@@ -399,14 +398,18 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
 
     _started = YES;
     // autoDetectErrors disables all unhandled event reporting
-    BOOL configuredToReportOOMs = self.configuration.reportOOMs && self.configuration.autoDetectErrors;
+    BOOL configuredToReportOOMs = self.configuration.autoDetectErrors
+        && ([[Bugsnag configuration] enabledErrorTypes] & BSGErrorTypesOOMs);
+    
     // Disable if a debugger is enabled, since the development cycle of starting
     // and restarting an app is also an uncatchable kill
     BOOL noDebuggerEnabled = !bsg_ksmachisBeingTraced();
+
     // Disable if in an app extension, since app extensions have a different
     // app lifecycle and the heuristic used for finding app terminations rooted
     // in fixable code does not apply
     BOOL notInAppExtension = ![BSG_KSSystemInfo isRunningInAppExtension];
+
     if (configuredToReportOOMs && noDebuggerEnabled && notInAppExtension) {
         [self.oomWatchdog enable];
     }
@@ -533,13 +536,11 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
         if (connected)
             [strongSelf flushPendingReports];
 
-        if (strongSelf.configuration.automaticallyCollectBreadcrumbs) {
-            [strongSelf addBreadcrumbWithBlock:^(BugsnagBreadcrumb *crumb) {
-                crumb.message = @"Connectivity change";
-                crumb.type = BSGBreadcrumbTypeState;
-                crumb.metadata  = @{ @"type"  :  connectionType };
-            }];
-        }
+        [strongSelf addBreadcrumbWithBlock:^(BugsnagBreadcrumb *crumb) {
+            crumb.message = @"Connectivity change";
+            crumb.type = BSGBreadcrumbTypeState;
+            crumb.metadata  = @{ @"type"  :  connectionType };
+        }];
     }];
 }
 
@@ -810,13 +811,6 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
     [[self state] addAttribute:BSGKeyOrientation
                      withValue:orientation
                  toTabWithName:BSGKeyDeviceState];
-    if ([self.configuration automaticallyCollectBreadcrumbs]) {
-        [self addBreadcrumbWithBlock:^(BugsnagBreadcrumb *_Nonnull breadcrumb) {
-          breadcrumb.type = BSGBreadcrumbTypeState;
-          breadcrumb.message = orientationNotifName;
-          breadcrumb.metadata = @{BSGKeyOrientation : orientation};
-        }];
-    }
 }
 
 - (void)lowMemoryWarning:(NSNotification *)notif {
@@ -824,9 +818,6 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
                      withValue:[[Bugsnag payloadDateFormatter]
                                    stringFromDate:[NSDate date]]
                  toTabWithName:BSGKeyDeviceState];
-    if ([self.configuration automaticallyCollectBreadcrumbs]) {
-        [self sendBreadcrumbForNotification:notif];
-    }
 }
 #endif
 
@@ -834,7 +825,7 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
     if (self.configuration.autoDetectErrors) {
         // Enable all crash detection
         bsg_kscrash_setHandlingCrashTypes(BSG_KSCrashTypeAll);
-        if (self.configuration.reportOOMs) {
+        if (self.configuration.enabledErrorTypes & BSGErrorTypesOOMs) {
             [self.oomWatchdog enable];
         }
     } else {
@@ -844,49 +835,6 @@ NSString *const BSGBreadcrumbLoadedMessage = @"Bugsnag loaded";
         [self.oomWatchdog disable];
     }
 }
-
-- (void)updateAutomaticBreadcrumbDetectionSettings {
-    if ([self.configuration automaticallyCollectBreadcrumbs]) {
-        for (NSString *name in [self automaticBreadcrumbStateEvents]) {
-            [self crumbleNotification:name];
-        }
-        for (NSString *name in [self automaticBreadcrumbTableItemEvents]) {
-            [[NSNotificationCenter defaultCenter]
-                addObserver:self
-                   selector:@selector(sendBreadcrumbForTableViewNotification:)
-                       name:name
-                     object:nil];
-        }
-        for (NSString *name in [self automaticBreadcrumbControlEvents]) {
-            [[NSNotificationCenter defaultCenter]
-                addObserver:self
-                   selector:@selector(sendBreadcrumbForControlNotification:)
-                       name:name
-                     object:nil];
-        }
-        for (NSString *name in [self automaticBreadcrumbMenuItemEvents]) {
-            [[NSNotificationCenter defaultCenter]
-                addObserver:self
-                   selector:@selector(sendBreadcrumbForMenuItemNotification:)
-                       name:name
-                     object:nil];
-        }
-    } else {
-        NSArray *eventNames = [[[[self automaticBreadcrumbStateEvents]
-            arrayByAddingObjectsFromArray:[self
-                                              automaticBreadcrumbControlEvents]]
-            arrayByAddingObjectsFromArray:
-                [self automaticBreadcrumbMenuItemEvents]]
-            arrayByAddingObjectsFromArray:
-                [self automaticBreadcrumbTableItemEvents]];
-        for (NSString *name in eventNames) {
-            [[NSNotificationCenter defaultCenter] removeObserver:self
-                                                            name:name
-                                                          object:nil];
-        }
-    }
-}
-
 - (NSArray<NSString *> *)automaticBreadcrumbStateEvents {
 #if TARGET_OS_TV
     return @[