+++ Tests/BugsnagSinkTests.m
@@ -136,7 +136,13 @@
     [self.rawReportData valueForKeyPath:@"user.state.crash.breadcrumbs"];
     NSArray *breadcrumbs =
     [self.processedData[@"events"] firstObject][@"breadcrumbs"];
-    XCTAssertEqualObjects(breadcrumbs, expected);
+    XCTAssertEqual(2, breadcrumbs.count);
+    for (int i = 0; i < breadcrumbs.count; i++) {
+        XCTAssertEqualObjects(expected[i][@"message"], breadcrumbs[i][@"message"]);
+        XCTAssertEqualObjects(expected[i][@"type"], breadcrumbs[i][@"type"]);
+        XCTAssertEqualObjects(expected[i][@"timestamp"], breadcrumbs[i][@"timestamp"]);
+        XCTAssertEqualObjects(expected[i][@"metadata"], breadcrumbs[i][@"metadata"]);
+    }
 }
 
 - (void)testEventContext {